Updates: - Added in any number of random number of jokes inside the request - Adjusted README.md
Technical Details:
The README.md was edited to include the underneath functionality.

I also added the functionality of specifying the random number of jokes you want. Unfortunately this functionality does not allow you to go to specify the type. This functionality was build using Express.js's   ability to match relative paths to wildcards. To get a specified random number of jokes navigate to the following relative path:
/jokes/random/{your number here}
Replace the {your number here} with your specified number. The number must be an integer and must not be over the the number of jokes or an error is returned. The original functionality of the default number    being 1 if there is no number specified is still there. To make this functionality I created a function in the handler.js which returns the random jokes given the num argument. The function was added into the   exports. It was also imported in the index.js. Next, I added a get method into the express application listening for a wildcard that represented by the above relative path. Next, I used substring to take in the number and parse it as an INT. This is where exceptional handling is used to return an error if the string can't be parsed. In the except clause, the error is returned. In the finally clause, it checks whether  the number is too long compared to the number of jokes and returns an error if so. If not then the application returns the jokes. The JSON of the jokes also had to be accessed by exporting it in the handler.js  and importing it in the index.js. It had to be accessed to get the number of jokes to see if a error has to be returned.